Moved insertonly classes in a folder
