+++ Source/EventFlow.MongoDB/ReadStores/InsertOnly/IMongoDbInsertOnlyReadModelStore.cs
@@ -1,6 +1,6 @@
 ﻿using EventFlow.ReadStores;
 
-namespace EventFlow.MongoDB.ReadStores
+namespace EventFlow.MongoDB.ReadStores.InsertOnly
 {
     public interface IMongoDbInsertOnlyReadModelStore<TReadModel> : IReadModelStore<TReadModel>
         where TReadModel : class, IReadModel, new()
